split asos from redcap
