+++ rp_asos/admin.py
@@ -1,0 +1,5 @@
+from django.contrib import admin
+from .models import Hospital, PatientRecord
+
+admin.site.register(Hospital)
+admin.site.register(PatientRecord)
