+++ rp_asos/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class RpAsosConfig(AppConfig):
+    name = "rp_asos"
