+++ rp_redcap/admin.py
@@ -1,8 +1,6 @@
 from django.contrib import admin
-from .models import Survey, Project, Contact, Hospital, PatientRecord
+from .models import Survey, Project, Contact
 
 admin.site.register(Project)
 admin.site.register(Survey)
 admin.site.register(Contact)
-admin.site.register(Hospital)
-admin.site.register(PatientRecord)
