+++ rp_redcap/views.py
@@ -2,11 +2,11 @@ from django.http import HttpResponse
 from rest_framework import status
 from rest_framework.views import APIView
 
-from .tasks import project_check, patient_data_check
+from .tasks import project_check
 from .models import Project
 
 
-def _validate_project_start_task(project_id, request, task):
+def validate_project_start_task(project_id, request, task):
     try:
         project = Project.objects.get(id=project_id)
     except Project.DoesNotExist:
@@ -22,17 +22,8 @@ def _validate_project_start_task(project_id, request, task):
 
 class StartProjectCheckView(APIView):
     def post(self, request, project_id, *args, **kwargs):
-        return_status = _validate_project_start_task(
+        return_status = validate_project_start_task(
             project_id, request, project_check
-        )
-
-        return HttpResponse(status=return_status)
-
-
-class StartPatientDataCheckView(APIView):
-    def post(self, request, project_id, *args, **kwargs):
-        return_status = _validate_project_start_task(
-            project_id, request, patient_data_check
         )
 
         return HttpResponse(status=return_status)
